+++ src/components/views/Board.vue
@@ -10,6 +10,9 @@
     <div v-if="!list.boardList" class="spinner">
       <img src="../../assets/spinner2.gif" alt="" />
     </div>
+    <h3 v-if="list.boardList < 1" class="no-results">
+      👆 Aun no tienes ninguna lista creada.
+    </h3>
     <div class="container">
       <Column
         class="list"
@@ -20,6 +23,7 @@
         @deleteBoardList="deleteBoardList"
         @addNewTask="addNewTask"
         @deleteNewTask="deleteNewTask"
+        @editNewTask="editNewTask"
       />
     </div>
   </section>
@@ -63,15 +67,17 @@ export default {
       this.listName = "";
     },
     async deleteBoardList(name) {
-      await this.deleteBoard(this.id);
-      const arr = this.list.boardList.filter((el) => el.name !== name);
-      const board = {
-        id: this.list.id,
-        name: this.list.name,
-        boardList: arr,
-      };
-      await this.postNewBoard(board);
-      await this.getList(this.id);
+      if (window.confirm("Do you really want to delete this list?")) {
+        await this.deleteBoard(this.id);
+        const arr = this.list.boardList.filter((el) => el.name !== name);
+        const board = {
+          id: this.list.id,
+          name: this.list.name,
+          boardList: arr,
+        };
+        await this.postNewBoard(board);
+        await this.getList(this.id);
+      }
     },
     async addNewTask(task) {
       await this.deleteBoard(this.id);
@@ -86,6 +92,22 @@ export default {
       await this.postNewBoard(board);
     },
     async deleteNewTask(obj) {
+      if (window.confirm("Do you really want to delete this task?")) {
+        await this.deleteBoard(this.id);
+        const board = {
+          id: this.list.id,
+          name: this.list.name,
+          boardList: this.list.boardList,
+        };
+        const arr = board.boardList
+          .find((el) => el.name === obj.name)
+          .taskList.filter((task) => task.task !== obj.task);
+
+        board.boardList.find((el) => el.name === obj.name).taskList = arr;
+        await this.postNewBoard(board);
+      }
+    },
+    async editNewTask(obj) {
       await this.deleteBoard(this.id);
       const board = {
         id: this.list.id,
@@ -94,9 +116,14 @@ export default {
       };
       const arr = board.boardList
         .find((el) => el.name === obj.name)
-        .taskList.filter((task) => task.task !== obj.task);
+        .taskList.filter((task) => task.task !== obj.task.task);
 
       board.boardList.find((el) => el.name === obj.name).taskList = arr;
+
+      board.boardList
+        .find((el) => el.name === obj.name)
+        .taskList.push(obj.task);
+
       await this.postNewBoard(board);
     },
   },
@@ -119,6 +146,11 @@ span {
 div.spinner img {
   margin-top: 50px;
   width: 100px;
+}
+
+h3.no-results {
+  color: whitesmoke;
+  margin-top: 50px;
 }
 
 div.container {
